+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Version: 1.0.19.c
 - fix voiding mode not working in quantum tanks
 - fix charger not dropping contents when broken
 - fix GT wrenches not being compatible with other mods
+- fix missing transfer size input in robot arms and fluid regulators, when a tag filter is used